test(backend): split rent/land tests into handler binding and service logic
Move the 22 existing cases verbatim to api/land_handler_test.go and
api/rent_handler_test.go, add a 422 binding case for estimate data
shortage, and add direct service tests covering stats, comparison,
theoretical-price estimation, appraisal sentinels, rent confidence
flags, and the rent query date range.
